patron_saga/app-commerce/requirements.txt
Flask==3.0.3
flask-marshmallow==1.2.1
marshmallow==3.23
requests==2.32.3
saga_py==1.1.0
Flask-Caching==2.3.0
redis==5.1.1
python-dotenv==1.0.1
Flask-Limiter==3.9.2
uwsgi==2.0.23